requirements.txt
python-dotenv
pandas
numpy
requests
langchain-community
langchain-openai
langchain
langgraph
openpyxl
pymongo
pymysql
pypdf
chromadb